Add Change::risk/describe & refactor migrations
Introduce Change::risk() and Change::describe() (returning ChangeDescription) and keep change_risk() delegating for compatibility. Move change presentation into the migrate crate and update the CLI TUI to use the new API. Refactor SchemaDiff into a DiffAccumulator with clearer domain passes and reorder pre/post type changes. Refactor DiffApplier to many focused sql_* helpers and add AddForeignKey helper. Export ChangeDescription from the crate. Also include various mechanical fixes: pattern simplifications in codegen, duration and sorting cleanups, inspector parsing fix, small DDL/test cleanups, and update CHANGELOG.

+++ crates/nautilus-engine/src/handlers/transactions.rs
@@ -19,7 +19,7 @@ pub(super) async fn handle_transaction_start(
 ) -> Result<Box<serde_json::value::RawValue>, ProtocolError> {
     let params: TransactionStartParams = parse_params(&request, "transactionStart")?;
 
-    let timeout = std::time::Duration::from_millis(params.timeout_ms.unwrap_or(5000) as u64);
+    let timeout = std::time::Duration::from_millis(params.timeout_ms.unwrap_or(5000));
 
     let tx_id = uuid::Uuid::new_v4().to_string();
     eprintln!("[engine] Starting transaction {}", tx_id);
@@ -82,7 +82,7 @@ pub(super) async fn handle_transaction_batch(
 ) -> Result<Box<serde_json::value::RawValue>, ProtocolError> {
     let params: TransactionBatchParams = parse_params(&request, "transactionBatch")?;
 
-    let timeout = std::time::Duration::from_millis(params.timeout_ms.unwrap_or(5000) as u64);
+    let timeout = std::time::Duration::from_millis(params.timeout_ms.unwrap_or(5000));
 
     let tx_id = uuid::Uuid::new_v4().to_string();
     eprintln!(
